api/pyproject.toml
[tool.poetry]
name = "photos-api"
version = "0.1.0"
description = ""
authors = ["Ryan Russell <rryanrussell@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.0"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
SQLAlchemy = "^1.4.22"
pydantic = "^1.8.2"
psycopg2-binary = "^2.9.1"
python-multipart = "^0.0.5"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
